1 try finished

+++ libft/ft_strtrim.c
@@ -6,15 +6,16 @@
 /*   By: dokwak <dokwak@student.42seoul.kr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/30 23:30:01 by dokwak            #+#    #+#             */
-/*   Updated: 2021/12/02 18:00:13 by dokwak           ###   ########.fr       */
+/*   Updated: 2021/12/03 18:01:21 by dokwak           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 #include "libft.h"
+/* size_t = -1 -> max size_t*/
 
 char	*ft_strtrim(char const *s1, char const *set)
 {
-	size_t	start;
-	size_t	end;
+	int		start;
+	int		end;
 	char	*pnt;
 
 	if (!s1 || !set)
